ElementSched -> ScheduleInfo, plus reduced idle counts from 32/8 -> 1

+++ elements/ip/setipdscp.cc
@@ -85,7 +85,7 @@ static String
 SetIPDSCP_read_dscp(Element *xf, void *)
 {
   SetIPDSCP *f = (SetIPDSCP *)xf;
-  return String(f->dscp()) + "\n";
+  return String((int)f->dscp()) + "\n";
 }
 
 void
